Fix spike smoothing
Spike smoothing was not working because the buffer would catch any extra
requests and then release them all at once without the protection of the
buffer.  This essentially created its own even more severe spike, and
only one of the released requests would be served.  By checking the
results of the released requests and rebuffering the failed ones, the
requests may take a little longer to resolve, but they only begin to
fail when the buffer is full.

In order to test spike smoothing the tests were updated to run the
requests asynchronously.